+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/storage/StorageCacheItem.java
@@ -11,13 +11,16 @@ import com.raoulvdberge.refinedstorage.apiimpl.API;
 import net.minecraft.item.ItemStack;
 
 import javax.annotation.Nonnull;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.function.BiConsumer;
 
 public class StorageCacheItem implements IStorageCache<ItemStack> {
     private INetwork network;
     private CopyOnWriteArrayList<IStorage<ItemStack>> storages = new CopyOnWriteArrayList<>();
     private IStackList<ItemStack> list = API.instance().createItemStackList();
+    private List<BiConsumer<ItemStack, Integer>> listeners = new LinkedList<>();
 
     public StorageCacheItem(INetwork network) {
         this.network = network;
@@ -56,6 +59,8 @@ public class StorageCacheItem implements IStorageCache<ItemStack> {
 
         if (!rebuilding) {
             network.sendItemStorageDeltaToClient(stack, size);
+
+            listeners.forEach(l -> l.accept(stack, size));
         }
     }
 
@@ -63,7 +68,14 @@ public class StorageCacheItem implements IStorageCache<ItemStack> {
     public synchronized void remove(@Nonnull ItemStack stack, int size) {
         if (list.remove(stack, size)) {
             network.sendItemStorageDeltaToClient(stack, -size);
+
+            listeners.forEach(l -> l.accept(stack, -size));
         }
+    }
+
+    @Override
+    public void addListener(BiConsumer<ItemStack, Integer> listener) {
+        listeners.add(listener);
     }
 
     @Override