+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.5.3
 - Fixed Solderer crashing (raoulvdberge)
+- The Portable Grid now exposes an inventory for interaction with other mods or vanilla (raoulvdberge)
 
 ### 1.5.2
 - Fixed a bug where loading nodes would abort when a single node has an error while reading (raoulvdberge)